// src/app2/modules/questionnaire/types/diligence-enum.type.ts
export enum DiligenceTypeFilters {
  dd_new = 'dd_new',
  dd_profile = 'dd_profile',
  dd_ongoing = 'dd_ongoing',
  dd_event_related = 'dd_event_related',
  dd_doc = 'dd_doc',
  dd_internal = 'dd_internal',
  shared_profile = 'shared_profile', // investor
  review_project_pre = 'review_project_pre',
  review_project_post = 'review_project_post',
  rating_project = 'rating_project',
  analyst_evaluation = 'analyst_evaluation',
}
export enum DiligenceTypeEnum {
  dd_review = 'dd_review', // analyst evaluation
  dd_profile = 'dd_profile', // preapproved
  dd_new = 'dd_new',
  dd_ongoing = 'dd_ongoing',
  dd_event_related = 'dd_event_related',
  dd_doc = 'dd_doc',
  inbound = 'inbound',
  shared_profile = 'shared_profile',
}